+++ src/utils.h
@@ -13,8 +13,10 @@ class Scaler
 private:
   rowvec z_mu, z_sd;
   bool standardize;
+  
 public:
   int n_col;
+  Scaler(); // Default constructor needed for serialization
   Scaler (mat z, bool standardize_);
   mat scale(mat z);
   mat unscale(mat z);
@@ -51,9 +53,12 @@ private:
   int k, curr_progress;
   double one_percent;
   std::string progressBar(int progress);
+  
 public:
   bool validate;
   int n_passes;
+  
+  Tracker(); // Default constructor needed for serialization
   Tracker(bool verbose_);
   mat train_history;
   void setTracker(int n_passes_, bool validate_, List train_param_);
